+++ Analysis/d2015_01_18_Hyperbolas/ex01--helloHyperbola.py
@@ -1,7 +1,6 @@
 import matplotlib.pyplot as plt
-import numpy as np
 
-from .common import get_focii
+from .common import get_focii, hyperbola, hyperbolaCOE
 
 # Initialize parameters
 d = 12.01e-2  # Hydrophone spacing (meters)
@@ -10,33 +9,6 @@ v = 1473  # Speed of sound in the medium (m/s)
 
 # Settings
 sz = "small"
-
-
-def hyperbola(a, b):
-    step = .1e-2
-    y = np.arange(-200, 200 + step, step)
-    x = a * np.sqrt(1 + (y / b)**2)
-    return (x, y)
-
-
-def hyperbolaCOE(c, del_t, v_medium):
-    # Compute a
-    del_d = del_t * v
-    a = del_d / 2
-
-    # check for error
-    if a > c:
-        print("hyperbolaCOE: ERROR. Your a, --%.1fcm--, is greater than c, --%.1fcm--. Please reduce time delay or increase element spacing."
-              % (a * 100, c * 100))
-        exit(1)
-
-    # Compute b
-    b = np.sqrt(c**2 - a**2)
-
-    # Explain and return
-    print("""hyperbolaCOE: I see you have a measured time delay of %.1f microseconds and the speed of sound in your medium is %d m/s. Given that your elements are %.1f cm appart, then a = %.1fcm, and b = %.1fcm."""
-          % (del_t * 1e6, v_medium, 2 * c * 100, a * 100, b * 100))
-    return (a, b)
 
 
 def main():
